Added Disable and Enable Timerjob to the sample
When trying to execute the sample, SharePoint 2016 complains that the timerjob should first be disabled before the StopAnalysis and ClearAnalysis commands are being run. Afterwards it requires running EnableTimerJobsSchedule to set the schedule back in place. I've updated the sample to include the disabling and enabling so the sample works.