Update permissions-in-the-security-and-compliance-center.md
It appears that the Security Readers role contains a copy and paste from the Security Administrators in the line "However, it will be a member of the Security Administrators role in Azure Active Directory and will inherit the capabilities of that role".  Granting someone Security Reader role in the Security and Compliance Center doesn't give them Security Administrator in Azure AD, or at least it shouldn't.